Externalize network map to powsybl-diagram-viewer (#1885)
