+++ package.json
@@ -12,7 +12,7 @@
         "@mui/icons-material": "^5.5.1",
         "@mui/lab": "^5.0.0-alpha.75",
         "@mui/material": "^5.5.3",
-        "@powsybl/diagram-viewer": "0.2.0",
+        "@powsybl/diagram-viewer": "0.3.1",
         "@reduxjs/toolkit": "^1.2.3",
         "@svgdotjs/svg.js": "^3.0.12",
         "@svgdotjs/svg.panzoom.js": "^2.1.2",
@@ -22,17 +22,12 @@
         "@types/react-dom": "^18.2.4",
         "ag-grid-community": "^29.0.0",
         "ag-grid-react": "^29.0.0",
-        "cheap-ruler": "^3.0.0",
         "clsx": "^1.0.4",
         "core-js": "^3.6.4",
         "dagre": "^0.8.5",
-        "deck.gl": "^8.9.18",
         "eventemitter3": "^5.0.0",
-        "geolib": "^3.2.1",
         "localized-countries": "^2.0.0",
         "lucene-escape-query": "^1.0.1",
-        "mapbox-gl": "^3.0.0",
-        "maplibre-gl": "^3.0.0",
         "mjolnir.js": "^2.4.0",
         "mui-nested-menu": "^3.0.0",
         "notistack": "^3.0.0",
@@ -47,7 +42,6 @@
         "react-hook-form": "^7.41.0",
         "react-intl": "^6.0.0",
         "react-json-view": "^1.21.3",
-        "react-map-gl": "^7.0.0",
         "react-papaparse": "^4.1.0",
         "react-plotly.js": "^2.6.0",
         "react-redux": "^8.0.0",
@@ -118,10 +112,10 @@
         "eslint-plugin-prettier": "^4.0.0",
         "eslint-plugin-react": "^7.21.0",
         "eslint-plugin-react-hooks": "^4.1.2",
+        "identity-obj-proxy": "^3.0.0",
         "jest": "^29.7.0",
         "jest-environment-jsdom": "^29.7.0",
         "jest-svg-transformer": "^1.0.0",
-        "identity-obj-proxy": "^3.0.0",
         "prettier": "^2.0.5",
         "ts-node": "^10.9.2",
         "vite": "^5.0.11",
